Documentation updated. Registration complete. OAuth exception added to registration endpoint

+++ rucsok-repository/src/main/java/com/rucsok/rucsok/repository/dao/RucsokRepository.java
@@ -19,7 +19,7 @@ public interface RucsokRepository extends CrudRepository<RucsokEntity, Long> {
 	public RucsokEntity findByLink(String link);
 	
 	@Query("SELECT r FROM Rucsok r WHERE id BETWEEN (:id -1) AND (:id + 1)")
-	public List<RucsokEntity> findById(@Param("id") long id);
+	public List<RucsokEntity> findById(@Param("id") int id);
 	
 	@Query("from Rucsok ORDER BY CreatedAt DESC")
 	public List<RucsokEntity> getAllRucsokByCreatedAt(Pageable pageable);
